feat: enable Auto Sync Permissions for Teams connector
- Add Teams to validAutoSyncSources frontend whitelist
- Fix parameter names in retrieve_all_slim_docs_perm_sync to match interface
  (_end -> end, _callback -> callback)

The Teams connector already implements SlimConnectorWithPermSync and
fetches external_access for each document, so only the frontend whitelist
and parameter name fix were needed to enable the feature.
