+++ backend/tests/daily/connectors/teams/test_teams_connector.py
@@ -1,10 +1,13 @@
 import os
 import time
+from collections.abc import Generator
 
 import pytest
 
 from onyx.access.models import ExternalAccess
+from onyx.connectors.models import Document
 from onyx.connectors.teams.connector import TeamsConnector
+from onyx.utils.variable_functionality import global_version
 from tests.daily.connectors.teams.models import TeamsThread
 from tests.daily.connectors.utils import load_everything_from_checkpoint_connector
 from tests.daily.connectors.utils import to_documents
@@ -166,3 +169,40 @@ def test_slim_docs_retrieval_from_teams_connector(
             slim_doc.external_access
         ), f"ExternalAccess should always be available, instead got {slim_doc=}"
         _assert_is_valid_external_access(external_access=slim_doc.external_access)
+
+
+@pytest.fixture(autouse=False)
+def set_ee_on() -> Generator[None, None, None]:
+    """Need EE to be enabled for perm sync tests to work since
+    perm syncing is an EE-only feature."""
+    global_version.set_ee()
+    yield
+    global_version._is_ee = False
+
+
+def test_load_from_checkpoint_with_perm_sync(
+    teams_connector: TeamsConnector,
+    set_ee_on: None,
+) -> None:
+    """Test that load_from_checkpoint_with_perm_sync returns documents with external_access.
+
+    This verifies the CheckpointedConnectorWithPermSync interface is properly implemented.
+    """
+    docs = load_everything_from_checkpoint_connector(
+        connector=teams_connector,
+        start=0.0,
+        end=time.time(),
+        include_permissions=True,  # Uses load_from_checkpoint_with_perm_sync
+    )
+
+    documents = [doc for doc in docs if isinstance(doc, Document)]
+
+    # We should have at least some documents
+    assert len(documents) > 0, "Expected to find at least one document"
+
+    for doc in documents:
+        assert (
+            doc.external_access is not None
+        ), f"Document {doc.id} should have external_access when using perm sync"
+
+        _assert_is_valid_external_access(external_access=doc.external_access)
